# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "基于 Go + MCP 协议的企业级零信任安全智能体平台 (AI SecOps)",
        "title": "ZeroTrust Sentinel API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Phantomor",
            "url": "https://github.com/Phantomor"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/chat": {
            "post": {
                "description": "处理基础的聊天请求，返回完整的字符串响应。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI 对话模块 (AI Chat)"
                ],
                "summary": "基础对话 (非流式)",
                "parameters": [
                    {
                        "description": "对话内容",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ChatRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 200, \"data\": \"AI回答的内容...\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "{\"error\": \"无效的请求参数\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "{\"error\": \"大脑宕机了: ...\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/chat/stream": {
            "post": {
                "description": "带有 RAG 知识库检索、上下文记忆、MySQL持久化的流式对话接口。使用 Server-Sent Events (SSE) 协议返回数据。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "AI 对话模块 (AI Chat)"
                ],
                "summary": "对话 (SSE 流式)",
                "parameters": [
                    {
                        "description": "包含 SessionID 的对话内容",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.StreamChatRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data: {\"content\": \"...\"}\\n\\ndata: [DONE]\\n\\n",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"error\": \"无效的请求参数\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "{\"error\": \"流式连接失败...\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/interview/report": {
            "post": {
                "description": "根据候选人历史面试记录、简历和 JD，调用大模型生成结构化的面试评估 JSON 报告。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AI 对话模块 (AI Chat)"
                ],
                "summary": "生成面试报告",
                "parameters": [
                    {
                        "description": "报告生成请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ReportRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": 0, \"data\": {\"score\": 85, \"strengths\": [...], \"hire_conclusion\": \"Hire\"}}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "{\"code\": 400, \"message\": \"参数错误或没有找到面试记录\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "{\"code\": 500, \"message\": \"生成报告失败\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/secops/compliance": {
            "post": {
                "description": "AI 合规顾问专用接口，基于企业零信任规范提供安全编码审查、漏洞修复建议与告警初步研判。具备独立的历史会话隔离 (Redis)。使用 Server-Sent Events (SSE) 协议实时返回审查分析结果。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "ZeroTrust AI 引擎"
                ],
                "summary": "DevSecOps 合规审查对话 (SSE 流式)",
                "parameters": [
                    {
                        "description": "包含 SessionID 的待审计代码片段、架构描述或告警日志",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.StreamChatRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data: {\\\"content\\\": \\\"...\\\"}\\n\\ndata: [DONE]\\n\\n",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\\\"error\\\": \\\"无效的安全审计请求参数\\\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "{\\\"error\\\": \\\"流式安全连接建立失败...\\\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/secops/threat_hunter": {
            "post": {
                "description": "基于 MCP 协议和 ReAct 引擎，提供自动化安全研判与本地沙箱执行能力",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "ZeroTrust AI 引擎"
                ],
                "summary": "ThreatHunter 核心调度接口",
                "parameters": [
                    {
                        "description": "智能体对话请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ThreatHunterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data: {\"content\": \"...\"}\\n\\ndata: [DONE]\\n\\n",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "{\"error\": \"参数错误\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "{\"error\": \"不支持流式或推理中断\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.ChatRequest": {
            "type": "object",
            "required": [
                "message"
            ],
            "properties": {
                "message": {
                    "type": "string",
                    "example": "什么是 GMP 模型？"
                }
            }
        },
...
        "dto.ReportRequest": {
            "type": "object",
            "required": [
                "jd",
                "resume",
                "session_id"
            ],
            "properties": {
                "jd": {
                    "type": "string",
                    "example": "需要3年Go开发经验，熟悉GMP模型..."
                },
                "resume": {
                    "type": "string",
                    "example": "熟练掌握Go语言，具有高并发经验..."
                },
                "session_id": {
                    "type": "string",
                    "example": "user_123_session"
                }
            }
        },
        "dto.StreamChatRequest": {
            "type": "object",
            "required": [
                "message",
                "session_id"
            ],
            "properties": {
                "message": {
                    "type": "string",
                    "example": "请解释一下 Go 的垃圾回收机制。"
                },
                "session_id": {
                    "type": "string",
                    "example": "user_123_session"
                }
            }
        },
        "dto.ThreatHunterRequest": {
            "type": "object",
            "required": [
                "message",
                "session_id"
            ],
            "properties": {
                "message": {
                    "type": "string",
                    "example": "帮我查一下这名候选人的背景信息"
                },
                "session_id": {
                    "type": "string",
                    "example": "threat_hunter_123_session"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}